Adjust PageViewController Scope
The contained PageViewController may, in some cases, need to be accessible externally. In my example, I need to mask the edges of the scroll view for a particular layout. Since this is an edge case that's layout-dependent, doesn't really make sense to add a full masking capability to Pageboy; rather, the simple solution is to provide access to the underlying controller.